fix(box-model):
coerce at least is a cheap shot,
the fundamental issue is that by the current model you measure left and rigt at the center of their respective columns, when they are the same the width is zero. the right should be considered to the right of the cell it designates, that way the width will be 1 when the start and the end column ( or row) have the same number.